+++ mlx/backend/no_cpu/CMakeLists.txt
@@ -1,0 +1,9 @@
+target_sources(
+  mlx
+  PRIVATE
+  ${CMAKE_CURRENT_SOURCE_DIR}/../common/load.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/primitives.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/../common/common.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/../common/compiled.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/../common/compiled_nocpu.cpp
+)
